Updated MultiChoiceQuestions stuff
